BD Fix

+++ backend/src/main/java/com/echocode/project/repositories/CategoryRepository.java
@@ -4,8 +4,8 @@ import com.echocode.project.entities.Category;
 import lombok.NonNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    public Category findByCategoryId(@NonNull int categoryId);
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    public Category findByCategoryId(@NonNull Long categoryId);
 
     public boolean existsByCategoryName(@NonNull String categoryName);
 }
